refactor(model): create models package
