+++ models/__init__.py
@@ -1,0 +1,5 @@
+""" models module init.py"""
+from .model import Model
+from .sir import SIR
+from .sirx import SIRX
+from .post_regression import ci_bootstrap, ci_block_cv
