+++ models/model.py
@@ -24,62 +24,6 @@ def call_solver(func, p, w0, t):
     return np.insert(sol, 0, t, axis=1)
 
 
-# pylint: disable=W0613
-def sir(w, t, p):
-    """ SIR: Simple model of disease spread
-    inputs:
-    w: vector of state variables [S,I,R]
-    where
-        S: Fraction of the population susceptible to the infection
-        I: Fraction on the population infected
-        R: Fraction of the population recovered
-    t: Current time
-    p: vector of parameter
-
-    returns:
-    f: right hand side of the system of differential equation
-    """
-    # unpack state variable
-    s, i, r = w  # pylint: disable=W0612
-    # unpack parameter
-    alpha, beta = p
-    ds_dt = -alpha * s * i
-    di_dt = alpha * s * i - beta * i
-    dr_dt = beta * i
-
-    return [ds_dt, di_dt, dr_dt]
-
-
-def sirx(w, t, p):
-    """ SIR-X: Dynamic outbreaks with temporally increasing
-    intervention
-
-    inputs:
-    w: vector of state variables [S,I,R,X]
-    where
-        S: Fraction of the population susceptible to the infection
-        I: Fraction on the population infected
-        R: Fraction of the population that recovered
-        X: Fraction of the population that is quarantined
-
-    t: time
-    p: vector of parameter
-
-    returns:
-    right hand side of the system of differential equation
-    """
-    # unpack state variable
-    s, i, r, x = w  # pylint: disable=W0612
-    # unpack parameter
-    alpha, beta, kappa_0, kappa = p
-    ds_dt = -alpha * s * i - kappa_0 * s
-    di_dt = alpha * s * i - beta * i - kappa_0 * i - kappa * i
-    dr_dt = kappa_0 * s + beta * i
-    dx_dt = (kappa_0 + kappa) * i
-
-    return [ds_dt, di_dt, dr_dt, dx_dt]
-
-
 class Model:
     """ Base model definition """
 
@@ -181,86 +125,18 @@ class Model:
         # Define fixed parameters: this set of parameters won't be fitted
         # fixed_params = self.p[fix_index]
 
-        def function_handle(t, *par_fit, population=population):
+        def function_handle(t, *par_fit, pop=population):
             params = np.array(self.p)
             params[fit_index] = par_fit
             self.p = params
             i_mod = call_solver(self.__class__.FUNC, self.p, self.w0, t)
-            return i_mod[:, 2] * population
+            return i_mod[:, 2] * pop
 
         # Fit parameters
         par_opt, pcov = curve_fit(
             f=function_handle, xdata=t_obs, ydata=n_i_obs, p0=fit_params0
         )
         self.p[fit_index] = par_opt
-        self.pcov = pcov
+        self.pcov = pcov  # This also flags that the model was fitted
         return self
         # return p_new, pcov
-
-
-class SIR(Model):
-    """ SIR model definition """
-
-    CSV_ROW = ["Days", "S", "I", "R"]
-    NUM_PARAMS = 2
-    NUM_IC = 3
-    FUNC = sir
-
-    def set_params(self, p, initial_conds):
-        """ Set model parameters.
-        input:
-        p: parameters of the model [alpha, beta]. All these
-           values should be in 1/day units.
-        initial_conds: Initial conditions (n_S0, n_I0, n_R0), where:
-          n_S0: Total number of susceptible to the infection
-          n_I0: Toral number of infected
-          n_R0: Total number of recovered
-          Note n_S0 + n_I0 + n_R0 = Population
-
-          Internally, the model initial conditions are the ratios
-          S0 = n_S0/Population
-          I0 = n_I0/Population
-          R0 = n_R0/Population
-          which is consistent with the mathematical description
-          of the SIR model.
-
-        output:
-        reference to self
-        """
-        self._set_params(p, initial_conds)
-        return self
-
-
-class SIRX(Model):
-    """ SIRX model definition """
-
-    CSV_ROW = ["Days", "S", "I", "R", "X"]
-    NUM_PARAMS = 4
-    NUM_IC = 4
-    FUNC = sirx
-
-    def set_params(self, p, initial_conds):
-        """ Set model parameters.
-        input:
-        p: parameters of the model [alpha, beta, kappa_0, kappa]. All these
-           values should be in 1/day units.
-        initial_conds: Initial conditions (S0, I0, R0, X0), where:
-          n_S0: Total number of susceptible to the infection
-          n_I0: Total number of infected
-          n_R0: Total number of recovered
-          n_X0: Total number of quarantined
-          Note: n_S0 + n_I0 + n_R0 + n_X0 = Population
-
-        Internally, the model initial conditions are the ratios
-          S0 = n_S0/Population
-          I0 = n_I0/Population
-          R0 = n_R0/Population
-          X0 = n_X0/Population
-          which is consistent with the mathematical description
-          of the SIR model.
-
-        output:
-        reference to self
-        """
-        self._set_params(p, initial_conds)
-        return self
